fix: fixing reviews count and rating isssues

+++ flutter_app/lib/data/models/serviceprovider_data.dart
@@ -2,7 +2,7 @@ class ServiceProvider {
   String name;
   String profession;
   String location;
-  double? rating;
+  int  ? rating;
   int? reviewCount;
   String jobsDone;
   String experience;
